+++ src/Notifications/MagicLinkNotification.php
@@ -3,6 +3,7 @@
 namespace Arzcode\FilamentMagicLogin\Notifications;
 
 use Arzcode\FilamentMagicLogin\Contracts\MagicLinkNotification as MagicLinkNotificationContract;
+use Arzcode\FilamentMagicLogin\Support\ExpiryDuration;
 use Illuminate\Bus\Queueable;
 use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
@@ -30,7 +31,9 @@ class MagicLinkNotification extends Notification implements MagicLinkNotificatio
         return (new MailMessage)
             ->subject(__('filament-magic-login::filament-magic-login.mail.subject', ['app' => config('app.name')]))
             ->greeting(__('filament-magic-login::filament-magic-login.mail.greeting'))
-            ->line(__('filament-magic-login::filament-magic-login.mail.intro', ['minutes' => $this->expiresAfterMinutes]))
+            ->line(__('filament-magic-login::filament-magic-login.mail.intro', [
+                'duration' => ExpiryDuration::describe($this->expiresAfterMinutes),
+            ]))
             ->action(__('filament-magic-login::filament-magic-login.mail.button'), $this->url)
             ->line(__('filament-magic-login::filament-magic-login.mail.ignore'))
             ->line(__('filament-magic-login::filament-magic-login.mail.fallback'))